# Conflicts:
#	src/main/java/ru/hogwarts/university/service/FacultyService.java
#	src/main/java/ru/hogwarts/university/service/StudentService.java

+++ src/main/java/ru/hogwarts/university/service/FacultyService.java
@@ -2,6 +2,7 @@ package ru.hogwarts.university.service;
 
 import org.springframework.stereotype.Service;
 import ru.hogwarts.university.exception.FacultyNotFoundException;
+import ru.hogwarts.university.exception.NameAndColourIsNullException;
 import ru.hogwarts.university.model.Faculty;
 import ru.hogwarts.university.repository.FacultyRepository;
 
@@ -27,8 +28,11 @@ public class FacultyService {
         return facultyRepository.save(faculty);
     }
 
-    public void deleteFaculty(Long facultyId) {
-        Faculty faculty = facultyRepository.findById(facultyId).orElseThrow(FacultyNotFoundException::new);
+    public void deleteFaculty(Long facultyId) { // existById
+        boolean facultyExists = facultyRepository.existsById(facultyId);
+        if (!facultyExists) {
+            throw new FacultyNotFoundException();
+        }
         facultyRepository.deleteById(facultyId);
     }
 
@@ -37,10 +41,19 @@ public class FacultyService {
     }
 
     public List<Faculty> getByColour(String colour) {
-        return facultyRepository.findByColour(colour);
+        return facultyRepository.findByColourIgnoreCase(colour);
     }
 
     public List<Faculty> findByNameOrColour(String name, String colour) {
+        if (name == null && colour != null) {
+            return facultyRepository.findByColourIgnoreCase(colour);
+        }
+        if (name != null && colour == null) {
+            return facultyRepository.findByNameIgnoreCase(name);
+        }
+        if (name == null && colour == null) {
+            throw new NameAndColourIsNullException();
+        }
         return facultyRepository.findByNameIgnoreCaseOrColourIgnoreCase(name, colour);
     }
 }
